combined funtofem_nlbgs_driver and tacs_interface for aerothermal/aeroelastic/aerothermoelastic cases
