+++ pyfuntofem/funtofem_nlbgs_driver.py
@@ -1,29 +1,36 @@
 #!/usr/bin/env python
+"""
+This file is part of the package FUNtoFEM for coupled aeroelastic simulation
+and design optimization.
 
-# This file is part of the package FUNtoFEM for coupled aeroelastic simulation
-# and design optimization.
+Copyright (C) 2015 Georgia Tech Research Corporation.
+Additional copyright (C) 2015 Kevin Jacobson, Jan Kiviaho and Graeme Kennedy.
+All rights reserved.
 
-# Copyright (C) 2015 Georgia Tech Research Corporation.
-# Additional copyright (C) 2015 Kevin Jacobson, Jan Kiviaho and Graeme Kennedy.
-# All rights reserved.
+FUNtoFEM is licensed under the Apache License, Version 2.0 (the "License");
+you may not use this software except in compliance with the License.
+You may obtain a copy of the License at
 
-# FUNtoFEM is licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this software except in compliance with the License.
-# You may obtain a copy of the License at
+   http://www.apache.org/licenses/LICENSE-2.0
 
-#   http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
 from __future__ import print_function
 
-from .funtofem_driver import *
+import numpy as np
+from mpi4py import MPI
+from funtofem import TransferScheme
+from .funtofem_driver import FUNtoFEMDriver
+try:
+    from .hermes_transfer import HermesTransfer
+except:
+    pass
 
-class FUNtoFEMnlbgs(FUNtoFEMDriver):
+class FUNtoFEMnlbgs_aerothermoelastic(FUNtoFEMDriver):
     def __init__(self,solvers,comm,struct_comm,struct_master,aero_comm,aero_master,transfer_options=None,model=None,
                  theta_init=0.125,theta_min=0.01,theta_max=1.0):
         """
@@ -45,8 +52,8 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
             Minimum value of theta for the Aitken under-relaxation
         """
 
-        super(FUNtoFEMnlbgs,self).__init__(solvers,comm,struct_comm,struct_master,aero_comm,aero_master,transfer_options=transfer_options,model=model)
-
+        super(FUNtoFEMnlbgs_aerothermoelastic,self).__init__(solvers,comm,struct_comm,struct_master,aero_comm,aero_master,transfer_options=transfer_options,model=model)
+        
         # Aitken acceleration settings
         self.theta_init = theta_init
         self.theta_min = theta_min
@@ -54,8 +61,11 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
         self.theta = []
 
         self.aitken_init = None
+        self.aitken_therm_init = None
         self.aitken_vec = None
+        self.aitken_therm_vec = None
         self.up_prev = None
+        self.therm_up_prev = None
 
     def _initialize_adjoint_variables(self,scenario,bodies):
         """
@@ -86,6 +96,21 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
             body.psi_D = np.zeros((body.aero_nnodes*3,nfunctions),
                                   dtype=TransferScheme.dtype)
 
+            # Thermal terms
+            body.psi_Q = np.zeros((body.struct_nnodes*body.therm_xfer_ndof,nfunctions),
+                                  dtype=TransferScheme.dtype)
+            body.psi_T_S = np.zeros((body.struct_nnodes*body.therm_xfer_ndof,nfunctions),
+                                  dtype=TransferScheme.dtype)
+            body.struct_rhs_T = np.zeros((body.struct_nnodes*body.therm_xfer_ndof,nfunctions),
+                                       dtype=TransferScheme.dtype)
+
+            body.dQdfta = np.zeros((body.aero_nnodes*4,nfunctions),
+                                  dtype=TransferScheme.dtype)
+            body.dAdta = np.zeros((body.aero_nnodes,nfunctions),
+                                  dtype=TransferScheme.dtype)
+            body.psi_T = np.zeros((body.aero_nnodes,nfunctions),
+                                  dtype=TransferScheme.dtype)
+
             if body.shape:
                 body.aero_shape_term = np.zeros((body.aero_nnodes*3,nfunctions_total),dtype=TransferScheme.dtype)
                 body.struct_shape_term = np.zeros((body.struct_nnodes*body.xfer_ndof,nfunctions_total),dtype=TransferScheme.dtype)
@@ -93,7 +118,7 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
 
     def _solve_steady_forward(self,scenario,steps=None):
         """
-        Solve the aeroelastic forward analysis using the nonlinear block Gauss-Seidel algorithm.
+        Solve the aerothermoelastic forward analysis using the nonlinear block Gauss-Seidel algorithm.
         Aitken under-relaxation for stabilty.
 
         Parameters
@@ -105,6 +130,7 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
         """
 
         self.aitken_init = True
+        self.aitken_therm_init = True
         fail = 0
 
         # Determine if we're using the scenario's number of steps or the argument
@@ -118,10 +144,16 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
 
         # Loop over the NLBGS steps
         for step in range(1,steps+1):
-            # Transfer displacements
+            # Transfer displacements and temperatures
             for body in self.model.bodies:
-                body.aero_disps = np.zeros(body.aero_nnodes*3,dtype=TransferScheme.dtype)
-                body.transfer.transferDisps(body.struct_disps, body.aero_disps)
+                
+                if body.analysis_type == 'aeroelastic' or body.analysis_type == 'aerothermoelastic':
+                    body.aero_disps = np.zeros(body.aero_nnodes*3,dtype=TransferScheme.dtype)
+                    body.transfer.transferDisps(body.struct_disps, body.aero_disps)
+                    
+                if body.analysis_type == 'aerothermal' or body.analysis_type == 'aerothermoelastic':
+                    body.aero_temps = np.zeros(body.aero_nnodes,dtype=TransferScheme.dtype)
+                    body.thermal_transfer.transferTemp(body.struct_temps, body.aero_temps)
 
             # Take a step in the flow solver
             fail = self.solvers['flow'].iterate(scenario,self.model.bodies,step)
@@ -132,10 +164,19 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     print('Flow solver returned fail flag')
                 return fail
 
-            # Transfer the loads
+            # Transfer the loads and heat flux
             for body in self.model.bodies:
-                body.struct_loads = np.zeros(body.struct_nnodes*body.xfer_ndof,dtype=TransferScheme.dtype)
-                body.transfer.transferLoads(body.aero_loads, body.struct_loads)
+                
+                if body.analysis_type == 'aeroelastic' or body.analysis_type == 'aerothermoelastic':
+                    body.struct_loads = np.zeros(body.struct_nnodes*body.xfer_ndof,dtype=TransferScheme.dtype)
+                    body.transfer.transferLoads(body.aero_loads, body.struct_loads)
+
+                if body.analysis_type == 'aerothermal' or body.analysis_type == 'aerothermoelastic':
+                    body.struct_heat_flux = np.zeros(body.struct_nnodes,dtype=TransferScheme.dtype)
+                    # FUN3D returns x,y,z, and magnitude of the surface normal heat flux
+                    # only need magnitude for TACS
+                    heat_flux_magnitude = body.aero_heat_flux[3::4].copy(order='C')
+                    body.thermal_transfer.transferFlux(heat_flux_magnitude, body.struct_heat_flux)
 
             # Take a step in the FEM model
             fail = self.solvers['structural'].iterate(scenario,self.model.bodies,step)
@@ -148,7 +189,6 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
 
             # Under-relaxation for solver stability
             self._aitken_relax()
-
 
         # end solve loop
         return fail
@@ -163,8 +203,10 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
         scenario: :class:`~scenario.Scenario`
             The current scenario
         """
+
         fail = 0
         self.aitken_init = True
+        self.aitken_therm_init = True
 
         # how many steps to take
         steps = scenario.steps
@@ -179,23 +221,30 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
 
         # Initialize the adjoint variables
         nfunctions = scenario.count_adjoint_functions()
-        self._initialize_adjoint_variables(scenario,self.model.bodies)
+        self._initialize_adjoint_variables(scenario, self.model.bodies)
 
         # loop over the adjoint NLBGS solver
-        for step in range(1,steps+1):
-            # Get force terms for the flow solver
+        for step in range(1, steps+1):
+            # Get force and heat flux terms for the flow solver
             for body in self.model.bodies:
                 for func in range(nfunctions):
-                    # 'Solve' for load transfer adjoint variables
-                    body.psi_L[:,func] = body.psi_S[:,func]
-
                     # Transform load transfer adjoint variables using transpose Jacobian from
-                    # funtofem: dLdfA^T * psi_L = dDdus * psi_L
+                    # funtofem: dLdfA^T * psi_L = dDdus * psi_S
                     psi_L_r = np.zeros(body.aero_nnodes*3,dtype=TransferScheme.dtype)
-                    body.transfer.applydDduS(body.psi_L[:, func].copy(order='C'), psi_L_r)
+                    body.transfer.applydDduS(body.psi_S[:, func].copy(order='C'), psi_L_r)
                     body.dLdfa[:,func] = psi_L_r
 
-            fail = self.solvers['flow'].iterate_adjoint(scenario,self.model.bodies,step)
+                    # Transform heat flux transfer adjoint variables using transpose Jacobian from
+                    # funtofem: dQdftA^T * psi_Q = dTdts * psi_Q
+                    psi_Q_r = np.zeros(body.aero_nnodes, dtype=TransferScheme.dtype)
+                    body.thermal_transfer.applydQdqATrans(body.psi_T_S[:, func].copy(order='C'), psi_Q_r)
+
+                    # Only set heat flux magnitude component of thermal adjoint in FUN3D
+                    # Can either use surface normal magnitude OR x,y,z components, not both
+                    body.dQdfta[3::4,func] = psi_Q_r
+
+            # Iterate over the aerodynamic adjoint
+            fail = self.solvers['flow'].iterate_adjoint(scenario, self.model.bodies, step)
 
             fail = self.comm.allreduce(fail)
             if fail != 0:
@@ -215,8 +264,14 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     # calculate dLdu_s^T * psi_L
                     psi_L_product = np.zeros(body.struct_nnodes*body.xfer_ndof,dtype=TransferScheme.dtype)
                     body.transfer.applydLduSTrans(body.psi_L[:, func].copy(order='C'), psi_L_product)
-
+                    # structural elastic rhs
                     body.struct_rhs[:,func] = -psi_D_product - psi_L_product
+
+                    # calculate dTdt_s^T * psi_T
+                    psi_T_product = np.zeros(body.struct_nnodes*body.therm_xfer_ndof, dtype=TransferScheme.dtype)
+                    body.psi_T = body.dAdta
+                    body.thermal_transfer.applydTdtSTrans(body.psi_T[:, func].copy(order='C'), psi_T_product)
+                    body.struct_rhs_T[:,func] = -psi_T_product
 
             # take a step in the structural adjoint
             fail = self.solvers['structural'].iterate_adjoint(scenario,self.model.bodies,step)
@@ -227,11 +282,11 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     print('Structural solver returned fail flag')
                 return fail
             self._aitken_adjoint_relax(scenario)
+            #self._aitken_adjoint_relax_thermal(scenario)
 
         # end of solve loop
 
         self._extract_coordinate_derivatives(scenario,self.model.bodies,steps)
-
         return 0
 
     def _solve_unsteady_forward(self,scenario,steps=None):
@@ -261,12 +316,15 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     print("No number of steps given for the coupled problem. Using default (1000)")
                 steps = 1000
 
-        for step in range(1,steps+1):
-            # Transfer structural displacements to aerodynamic surface
+        for step in range(1, steps+1):
+            # Transfer structural displacements and temperatures to aerodynamic surface
             for body in self.model.bodies:
 
                 body.aero_disps = np.zeros(body.aero_nnodes*3,dtype=TransferScheme.dtype)
                 body.transfer.transferDisps(body.struct_disps, body.aero_disps)
+
+                body.aero_temps = np.zeros(body.aero_nnodes, dtype=TransferScheme.dtype)
+                body.thermal_transfer.transferTemp(body.struct_temps, body.aero_temps)
 
                 if ('rigid'  in body.motion_type and
                     'deform' in body.motion_type):
@@ -286,7 +344,8 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                 elif('rigid' in body.motion_type):
                     transform = self.solvers['structural'].get_rigid_transform(body)
 
-            fail = self.solvers['flow'].iterate(scenario,self.model.bodies,step)
+            # Take a time step for the flow solver
+            fail = self.solvers['flow'].iterate(scenario, self.model.bodies, step)
 
             fail = self.comm.allreduce(fail)
             if fail != 0:
@@ -294,13 +353,17 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     print('Flow solver returned fail flag')
                 return fail
 
-            # Transfer loads from fluid and get loads on structure
+            # Transfer loads and heat flux from fluid and get loads and temps on structure
             for body in self.model.bodies:
                 body.struct_loads = np.zeros(body.struct_nnodes*body.xfer_ndof, dtype=TransferScheme.dtype)
                 body.transfer.transferLoads(body.aero_loads, body.struct_loads)
 
+                body.struct_heat_flux = np.zeros(body.struct_nnodes, dtype=TransferScheme.dtype)
+                heat_flux_magnitude = body.aero_heat_flux[3::4].copy(order='C')
+                body.thermal_transfer.transferFlux(heat_flux_magnitude, body.struct_heat_flux)
+
             # Take a step in the FEM model
-            fail = self.solvers['structural'].iterate(scenario,self.model.bodies,step)
+            fail = self.solvers['structural'].iterate(scenario, self.model.bodies, step)
 
             fail = self.comm.allreduce(fail)
             if fail != 0:
@@ -309,6 +372,7 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                 return fail
 
         # end solve loop
+
         return fail
 
     def _solve_unsteady_adjoint(self,scenario):
@@ -328,17 +392,22 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
             fail flag
 
         """
-        # Initialize the adjoint variables
-        nfunctions = scenario.count_adjoint_functions()
-        self._initialize_adjoint_variables(scenario,self.model.bodies)
 
+        fail = 0
+
+        # how many steps to take
         steps = scenario.steps
 
+        # Initialize the adjoint variables
+        nfunctions = scenario.count_adjoint_functions()
+        self._initialize_adjoint_variables(scenario, self.model.bodies)
+
+        # Loop over each time step in the reverse order
         for rstep in range(1,steps+1):
             step = steps - rstep + 1
 
             self.solvers['flow'].set_states(scenario,self.model.bodies,step)
-            # Due to the staggering, we linearize the transfer about u_s^(n-1)
+            # Due to the staggering, we linearize the transfer about t_s^(n-1)
             self.solvers['structural'].set_states(scenario,self.model.bodies,step-1)
 
             for body in self.model.bodies:
@@ -366,7 +435,7 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     body.aero_disps = u.copy()
 
             # take a step in the structural adjoint
-            fail = self.solvers['structural'].iterate_adjoint(scenario,self.model.bodies,step)
+            fail = self.solvers['structural'].iterate_adjoint(scenario, self.model.bodies, step)
 
             fail = self.comm.allreduce(fail)
             if fail != 0:
@@ -374,18 +443,22 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     print('Structural solver returned fail flag')
                 return fail
 
+            # Get load and heat flux terms for the flow solver
             for body in self.model.bodies:
                 for func in range(nfunctions):
-                    # 'Solve' for load transfer adjoint variables
-                    body.psi_L[:,func] = body.psi_S[:,func]
-
                     # Transform load transfer adjoint variables using transpose Jacobian from
                     # funtofem: dLdfA^T * psi_L
                     psi_L_r = np.zeros(body.aero_nnodes*3,dtype=TransferScheme.dtype)
-                    body.transfer.applydDduS(body.psi_L[:, func].copy(order='C'), psi_L_r)
+                    body.transfer.applydDduS(body.psi_S[:, func].copy(order='C'), psi_L_r)
                     body.dLdfa[:,func] = psi_L_r
 
-            fail = self.solvers['flow'].iterate_adjoint(scenario,self.model.bodies,step)
+                    # Transform heat flux transfer adjoint variables using transpose Jacobian from
+                    # funtofem: dQdftA^T * psi_Q = dTdts * psi_Q
+                    psi_Q_r = np.zeros(body.aero_nnodes, dtype=TransferScheme.dtype)
+                    body.thermal_transfer.applydQdqATrans(body.psi_T_S[:, func].copy(order='C'), psi_Q_r)
+                    body.dQdfta[3::4,func] = psi_Q_r
+
+            fail = self.solvers['flow'].iterate_adjoint(scenario, self.model.bodies, step)
 
             fail = self.comm.allreduce(fail)
             if fail != 0:
@@ -449,6 +522,12 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     body.transfer.applydLduSTrans(body.psi_L[:,func].copy(order='C'), psi_L_product)
                     body.struct_rhs[:,func] = -psi_D_product - psi_L_product
 
+                    # calculate dTdt_s^T * psi_T
+                    psi_T_product = np.zeros(body.struct_nnodes*body.therm_xfer_ndof, dtype=TransferScheme.dtype)
+                    body.psi_T = body.dAdta
+                    body.thermal_transfer.applydTdtSTrans(body.psi_T[:, func].copy(order='C'), psi_T_product)
+                    body.struct_rhs_T[:,func] = -psi_T_product
+
             # extract and accumulate coordinate derivative every step
             self._extract_coordinate_derivatives(scenario,self.model.bodies,step)
 
@@ -461,6 +540,7 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
             if self.comm.Get_rank() == 0:
                 print('Flow solver returned fail flag')
             return fail
+
         fail = self.solvers['structural'].iterate_adjoint(scenario, self.model.bodies, step=0)
         fail = self.comm.allreduce(fail)
         if fail != 0:
@@ -469,12 +549,17 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
             return fail
 
         # extract coordinate derivative term from initial condition
-        self._extract_coordinate_derivatives(scenario,self.model.bodies,step=0)
+        self._extract_coordinate_derivatives(scenario, self.model.bodies, step=0)
 
-        fail = 0
-        return fail
+        return 0
 
     def _aitken_relax(self):
+        """
+        Include documentation - follow original format^
+        """
+
+
+
         if self.aitken_init:
             self.aitken_init = False
 
@@ -506,7 +591,6 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                 self.aitken_vec[ibody] += self.theta[ibody] * up
                 self.up_prev[ibody] = up[:]
                 body.struct_disps = self.aitken_vec[ibody]
-
         return
 
     def _aitken_adjoint_relax(self,scenario):
@@ -545,5 +629,77 @@ class FUNtoFEMnlbgs(FUNtoFEMDriver):
                     self.aitken_vec[ibody][func] += self.theta[ibody][func] * up
                     self.up_prev[ibody][func] =up[:]
                     body.psi_S[:,func] = self.aitken_vec[ibody][func][:]
-
         return self.aitken_vec
+
+
+    def _aitken_relax_thermal(self):
+        if self.aitken_therm_init:
+            self.aitken_therm_init = False
+
+            # initialize the 'previous update' to zero
+            self.therm_up_prev = []
+            self.aitken_therm_vec = []
+            self.theta = []
+
+            for ind, body in enumerate(self.model.bodies):
+                self.therm_up_prev.append(body.T_ref*np.ones(body.struct_nnodes*body.therm_xfer_ndof,dtype=TransferScheme.dtype))
+                self.aitken_therm_vec.append(body.T_ref*np.ones(body.struct_nnodes*body.therm_xfer_ndof,dtype=TransferScheme.dtype))
+                self.theta.append(self.theta_init)
+
+        # do the Aitken update
+        for ibody, body in enumerate(self.model.bodies):
+            if body.struct_nnodes > 0:
+                up = body.struct_temps - self.aitken_therm_vec[ibody]
+                norm2 = (np.linalg.norm(up - self.therm_up_prev[ibody])**2.0)
+
+                # Only update theta if the displacements changed
+                if norm2 > 1e-13:
+                    self.theta[ibody] *= 1.0 - (up - self.therm_up_prev[ibody]).dot(up)/norm2
+                    self.theta[ibody] = np.max((np.min((self.theta[ibody],self.theta_max)),self.theta_min))
+
+                # handle the min/max for complex step
+                if type(self.theta[ibody]) == np.complex128 or type(self.theta[ibody]) == complex:
+                    self.theta[ibody] = self.theta[ibody].real + 0.0j
+
+                self.aitken_therm_vec[ibody] += self.theta[ibody] * up
+                self.therm_up_prev[ibody] = up[:]
+                body.struct_temps = self.aitken_therm_vec[ibody]
+        return
+
+    def _aitken_adjoint_relax_thermal(self,scenario):
+        nfunctions =  scenario.count_adjoint_functions()
+        if self.aitken_therm_init:
+            self.aitken_therm_init = False
+
+            # initialize the 'previous update' to zero
+            self.therm_up_prev = []
+            self.aitken_therm_vec = []
+            self.theta = []
+
+            for ibody, body in enumerate(self.model.bodies):
+                up_prev_body = []
+                aitken_therm_vec_body = []
+                theta_body = []
+                for func in range(nfunctions):
+                    up_prev_body.append(body.T_ref*np.ones(body.struct_nnodes*body.therm_xfer_ndof,dtype=TransferScheme.dtype))
+                    aitken_therm_vec_body.append(body.T_ref*np.ones(body.struct_nnodes*body.therm_xfer_ndof,dtype=TransferScheme.dtype))
+                    theta_body.append(self.theta_init)
+                self.therm_up_prev.append(up_prev_body)
+                self.aitken_therm_vec.append(aitken_therm_vec_body)
+                self.theta.append(theta_body)
+
+        # do the Aitken update
+        for ibody, body in enumerate(self.model.bodies):
+            if body.struct_nnodes > 0:
+                for func in range(nfunctions):
+                    up = body.psi_T_S[:,func] - self.aitken_therm_vec[ibody][func]
+                    norm2 = np.linalg.norm(up - self.therm_up_prev[ibody][func])**2.0
+
+                    # Only update theta if the vector changed
+                    if norm2 > 1e-13:
+                        self.theta[ibody][func] *= 1.0 - (up - self.therm_up_prev[ibody][func]).dot(up)/np.linalg.norm(up - self.therm_up_prev[ibody][func])**2.0
+                        self.theta[ibody][func] = np.max((np.min((self.theta[ibody][func],self.theta_max)),self.theta_min))
+                    self.aitken_therm_vec[ibody][func] += self.theta[ibody][func] * up
+                    self.therm_up_prev[ibody][func] =up[:]
+                    body.psi_T_S[:,func] = self.aitken_therm_vec[ibody][func][:]
+        return self.aitken_therm_vec
